refactor(signup): 공개 라우트 sign-up 엔트리 배럴 경계 분리 (#480)
