+++ app/(public)/sign-up/intro/page.tsx
@@ -1,4 +1,4 @@
-import { SignUpIntroPage } from "@/pages/auth";
+import { SignUpIntroPage } from "@/pages/sign-up";
 
 export default function Page() {
   return <SignUpIntroPage />;
