+++ src/pages/sign-up/index.ts
@@ -1,0 +1,1 @@
+export { SignUpPage, SignUpIntroPage } from "../auth/sign-up";
